Refactoring code and connect to CLI
